updated iliad

+++ src/inputs/position.rs
@@ -2,7 +2,7 @@ use serde::Deserialize;
 
 #[derive(Deserialize, Debug)]
 pub struct InputPositionUpdate {
-    pub chapter_index: u32,
-    pub chapter_position: u64,
+    pub chapter_index: i64,
+    pub chapter_position: i64,
     pub client_timestamp: i64,
 }
