+++ src/models/position.rs
@@ -2,10 +2,11 @@ use chrono::NaiveDateTime;
 use sqlx::FromRow;
 
 #[derive(Debug, FromRow)]
+#[allow(dead_code)]
 pub struct Position {
     pub audiobook_hash: String, // primary key
     pub username: String,       // primary key
-    pub chapter_index: u32,
-    pub chapter_position: u64,
+    pub chapter_index: i64,
+    pub chapter_position: i64,
     pub timestamp: NaiveDateTime,
 }
