+++ src/outputs/position.rs
@@ -2,6 +2,6 @@ use serde::Serialize;
 
 #[derive(Serialize, Debug)]
 pub struct OutputPositionUpdate {
-    pub chapter_index: u32,
-    pub chapter_position: u64,
+    pub chapter_index: i64,
+    pub chapter_position: i64,
 }
